implemented display form and entry details

+++ src/components/Details.js
@@ -8,16 +8,15 @@ function Details(props) {
 
     const [formLabels, setformLabels] = useState({
         heading: '',
-        org: ''
+        orgLabel: ''
     })
 
-    const [formDetails, setFormDetails] = useState({
-        org: '',
-        title: '',
-        startDate: '',
-        endDate: '',
-        details: ''
-    })
+    const [edit, setEdit] = useState(false);
+    const [org, setOrg] = useState('');
+    const [title, setTitle] = useState('');
+    const [startDate, setStartDate] = useState('');
+    const [endDate, setEndDate] = useState('');
+    const [details, setDetails] = useState('');
 
     const [entries, setEntries] = useState(() => {
 
@@ -29,15 +28,26 @@ function Details(props) {
 
     })
 
-    function showForm(heading, org) {
+    function showForm(heading, orgLabel) {
 
         setformLabels({
             heading: heading,
-            org: org
+            orgLabel: orgLabel
         })
 
         props.toggleHidden();
 
+    }
+
+    function closeForm() {
+
+        setOrg('');
+        setTitle('');
+        setStartDate('');
+        setEndDate('');
+        setDetails('');
+
+        props.toggleHidden();
     }
 
     function addEntry(entry) {
@@ -52,34 +62,86 @@ function Details(props) {
 
     function deleteEntry(entryId) {
 
-        entries.delete(entryId)
-        const updatedEntries = new Map(entries);
+        if (window.confirm('Are you sure?')) {
 
-        localStorage.setItem('entries', JSON.stringify(Object.fromEntries(updatedEntries)));
-        setEntries(updatedEntries);
+            entries.delete(entryId)
+            const updatedEntries = new Map(entries);
+
+            localStorage.setItem('entries', JSON.stringify(Object.fromEntries(updatedEntries)));
+            setEntries(updatedEntries);
+
+        }
 
     }
 
     // TODO: implement edit functionality
-    function editEntry(entryId, section, org) {
+    function editEntry(entryId, section, orgLabel) {
         
-        showForm(section,org);
+        setEdit(true);
 
+        // populate details
+        let entry = entries.get(entryId);
+        setOrg(entry.org);
+        setTitle(entry.title);
+        setStartDate(entry.startDate);
+        setEndDate(entry.endDate);
+        setDetails(entry.details);
+
+        // change 'Add' to 'Save changes'
+
+        showForm(section,orgLabel);
+
+    }
+
+    function handleChange(event) {
+
+        switch(event.target.name) {
+            case 'Institution' || 'Company' || 'Organisation':
+                setOrg(event.target.value);
+                break;
+            case 'Title':
+                setTitle(event.target.value);
+                break;
+            case 'Start Date':
+                setStartDate(event.target.value);
+                break;
+            case 'End Date':
+                setEndDate(event.target.value);
+                break;
+            case 'Details':
+                setDetails(event.target.value);
+                break;
+            default:
+        }
 
     }
 
     return (
         <div className="container flex flex-col flex-ai-c flex-jc-se">
-            {props.hidden && <Form add={addEntry} close={props.toggleHidden} heading={formLabels.heading} org={formLabels.org} />}
+            {props.hidden && 
+                <Form 
+                    add={addEntry} 
+                    close={closeForm} 
+                    edit={edit}
+                    heading={formLabels.heading} 
+                    orgLabel={formLabels.orgLabel} 
+                    change={handleChange}
+                    org={org}
+                    title={title}
+                    startDate={startDate}
+                    endDate={endDate}
+                    details={details}
+                />
+            }
             <Heading text="Make-a-CV"/>
             <div className="input flex flex-col flex-jc-sb">
                 <InputField label="Name" />
                 <InputField label="Phone" />
                 <InputField label="Email" />
             </div>
-            <Section showForm={showForm} edit={editEntry} delete={deleteEntry} entries={entries} heading="Education" org="Institution" />
-            <Section showForm={showForm} edit={editEntry} delete={deleteEntry} entries={entries} heading="Work Experience" org="Company" />
-            <Section showForm={showForm} edit={editEntry} delete={deleteEntry} entries={entries} heading="Extra-curricular" org="Organisation" />
+            <Section showForm={showForm} edit={editEntry} delete={deleteEntry} entries={entries} heading="Education" orgLabel="Institution" />
+            <Section showForm={showForm} edit={editEntry} delete={deleteEntry} entries={entries} heading="Work Experience" orgLabel="Company" />
+            <Section showForm={showForm} edit={editEntry} delete={deleteEntry} entries={entries} heading="Extra-curricular" orgLabel="Organisation" />
         </div>
     )
 }
